Bumping version. Making GCC 4.3.4 happy.

+++ edit.c
@@ -153,7 +153,7 @@ static int deleteSong(char *args, void *data){
 	char query[100],*ptr;
 	int x;
 	printf("Delete?: ");
-	fgets(args,200,stdin);
+	if(!fgets(args,200,stdin))return 1;
 	if(*args=='y' || *args=='Y'){
 		for(x=0;x<songids->length;x++){
 			sprintf(query,"DELETE FROM Song WHERE SongID=%d",songids->songid[x]);
@@ -318,7 +318,7 @@ static int deletePlaylist(char *args, void *data){
 	int x;
 
 	printf("Delete playlist?: ");
-	fgets(args,200,stdin);
+	if(!fgets(args,200,stdin))return 1;
 	if(*args=='y' || *args=='Y'){
 		for(x=0;x<ids->length;x++){
 			sprintf(query,"DELETE FROM PlaylistSong WHERE PlaylistID=%d",ids->songid[x]);
@@ -403,8 +403,7 @@ static void getPlaylistSongOrders(char *args, int *cur, int*new){
 	for(x=1;args[x] && (args[x]<'0' || args[x]>'9');x++);
 	if(!args[x]){
 		printf("Current order: ");
-		fgets(args,200,stdin);
-		if(*args=='\n'){
+		if(!fgets(args,200,stdin) || *args=='\n'){
 			printf("Aborted\n");
 			*cur=*new=0;
 			return;
@@ -413,8 +412,7 @@ static void getPlaylistSongOrders(char *args, int *cur, int*new){
 		*cur=(int)strtol(args,NULL,10);
 
 		printf("New order: ");
-		fgets(args,200,stdin);
-		if(*args=='\n'){
+		if(!fgets(args,200,stdin) || *args=='\n'){
 			printf("Aborted\n");
 			*cur=*new=0;
 			return;
@@ -431,8 +429,7 @@ static void getPlaylistSongOrders(char *args, int *cur, int*new){
 		for(;args[x] && (args[x]<'0' || args[x]>'9');x++);
 		if(!args[x]){
 			printf("New order: ");
-			fgets(args,200,stdin);
-			if(*args=='\n'){
+			if(!fgets(args,200,stdin) || *args=='\n'){
 				printf("Aborted\n");
 				*cur=*new=0;
 				return;
@@ -564,7 +561,7 @@ static int editGenreParent(char *args, void *data){
 
 	for(x=0;x<ids->length;x++){
 		if(gid==ids->songid[x])continue;
-		sprintf(query,"UPDATE Category SET ParentID=%d WHERE CategoryID=%d AND CategoryID NOT IN (SELECT ParentID FROM Category WHERE CategoryID=%1$d)",gid,ids->songid[x]);
+		sprintf(query,"UPDATE Category SET ParentID=%1$d WHERE CategoryID=%2$d AND CategoryID NOT IN (SELECT ParentID FROM Category WHERE CategoryID=%1$d)",gid,ids->songid[x]);
 		debug(3,query);
 		sqlite3_exec(conn,query,NULL,NULL,NULL);
 	}
@@ -585,7 +582,7 @@ static int editGenreDelete(char *args, void *data){
 	for(x=0;x<ids->length;x++){
 		if(ids->songid[x]==1)continue;
 
-		sprintf(query,"UPDATE Category SET ParentID=(SELECT ParentID FROM Category WHERE CategoryID=%d) WHERE ParentID=%1$d",ids->songid[x]);
+		sprintf(query,"UPDATE Category SET ParentID=(SELECT ParentID FROM Category WHERE CategoryID=%1$d) WHERE ParentID=%1$d",ids->songid[x]);
 		sqlite3_exec(conn,query,NULL,NULL,NULL);
 		sprintf(query,"DELETE FROM Category WHERE CategoryID=%d",ids->songid[x]);
 		sqlite3_exec(conn,query,NULL,NULL,NULL);
@@ -615,7 +612,7 @@ static int editSongGenreAdd(char *args, void *data){
 	}
 
 	// Insert songs from list that are not already in the category
-	sprintf(query,"INSERT INTO SongCategory(CategoryID,SongID) SELECT '%d',SongID FROM SongCategory WHERE SongID NOT IN (SELECT SongID FROM SongCategory WHERE CategoryID=%1$d) AND SongID IN (SELECT SelectID FROM TempSelect WHERE TempID=%d)",gid,ids->tempselectid);
+	sprintf(query,"INSERT INTO SongCategory(CategoryID,SongID) SELECT '%1$d',SongID FROM SongCategory WHERE SongID NOT IN (SELECT SongID FROM SongCategory WHERE CategoryID=%1$d) AND SongID IN (SELECT SelectID FROM TempSelect WHERE TempID=%2$d)",gid,ids->tempselectid);
 	debug(3,query);
 	sqlite3_exec(conn,query,NULL,NULL,NULL);
 	return 1;
